+++ sites/partners/src/listings/PaperListingForm/sections/BuildingFeatures.tsx
@@ -1,12 +1,27 @@
-import React from "react"
+import React, { useMemo } from "react"
 import { useFormContext } from "react-hook-form"
-import { t, GridSection, Textarea } from "@bloom-housing/ui-components"
+import { t, GridSection, Textarea, ViewItem, FieldGroup } from "@bloom-housing/ui-components"
+import { listingFeatures } from "@bloom-housing/shared-helpers"
+import { ListingFeatures } from "@bloom-housing/backend-core/types"
 
-const BuildingFeatures = () => {
+type BuildingFeaturesProps = {
+  existingFeatures: ListingFeatures
+}
+
+const BuildingFeatures = (props: BuildingFeaturesProps) => {
   const formMethods = useFormContext()
 
   // eslint-disable-next-line @typescript-eslint/unbound-method
   const { register } = formMethods
+
+  const featureOptions = useMemo(() => {
+    return listingFeatures.map((item) => ({
+      id: item,
+      label: t(`eligibility.accessibility.${item}`),
+      defaultChecked: props.existingFeatures ? props.existingFeatures[item] : false,
+      register,
+    }))
+  }, [register])
 
   return (
     <div>
@@ -26,7 +41,7 @@ const BuildingFeatures = () => {
             maxLength={600}
           />
           <Textarea
-            label={t("t.accessibility")}
+            label={t("t.additionalAccessibility")}
             name={"accessibility"}
             id={"accessibility"}
             fullWidth={true}
@@ -69,6 +84,17 @@ const BuildingFeatures = () => {
             register={register}
             maxLength={600}
           />
+        </GridSection>
+        <GridSection columns={1}>
+          <ViewItem label={t("listings.sections.accessibilityFeatures")}>
+            <FieldGroup
+              type="checkbox"
+              name="listingFeatures"
+              fields={featureOptions}
+              register={register}
+              fieldGroupClassName="grid grid-cols-3 mt-4"
+            />
+          </ViewItem>
         </GridSection>
       </GridSection>
     </div>